refactor: centralize YouTube helpers
