+++ js/project1.js
@@ -13,6 +13,7 @@ import {
   createDesignProcessSection,
   createDetailsSection,
 } from "./layout.js";
+import { convertYouTubeUrl, parseYouTubeStartTime } from "./youtubeUtils.js";
 
 document.addEventListener("DOMContentLoaded", async () => {
   await loadHeader();
@@ -122,28 +123,4 @@ function renderSections() {
       container.appendChild(el);
     }
   });
-}
-
-function convertYouTubeUrl(url) {
-  const urlObj = new URL(url);
-  const videoId = urlObj.searchParams.get("v");
-  const start = urlObj.searchParams.get("t");
-
-  let embedUrl = `https://www.youtube.com/embed/${videoId}`;
-  if (start) {
-    // YouTube erlaubt Startzeit im Format Sekunden
-    const seconds = parseYouTubeStartTime(start);
-    embedUrl += `?start=${seconds}`;
-  }
-
-  return embedUrl;
-}
-
-function parseYouTubeStartTime(t) {
-  // "2s" oder "1m30s" -> Sekunden umrechnen
-  const match = t.match(/(?:(\d+)m)?(\d+)s/);
-  if (!match) return 0;
-  const minutes = parseInt(match[1] || 0, 10);
-  const seconds = parseInt(match[2] || 0, 10);
-  return minutes * 60 + seconds;
 }
